requirements.txt
tensorflow>=2.2.0
tqdm~=4.48.0
librosa~=0.8.0
numpy~=1.19.0
matplotlib~=3.3.0
scikit-learn~=0.23.1
PyYAML